Refactoring Warning/Error system (more than expected), new option '--msg=none'
* new option `--msg=none`
* Error(..)/Warning(..) types refactored and cleaned up + documented
behaviour
* LabelTable items have now also variable keeping number of pass when
they were updated at last. This is used to silently update value of
labels in pass 2 (instead of reporting "error: duplicate label")
* updated two tests to reflect current state + extend them

The label table pass-logic is more like quick hack to get desired result
on limited amount of ASM sources, and it may be fatally wrong or
incomplete, but so was the old code.

Surely add github issue in case it breaks your old source!

(this part of sjasmplus needs rather complete rewrite from scratch, but
that needs more tests and Documentation cleanup+additions to make the
desired behaviour well defined first - currently the docs are not clear
about what is "correct" behaviour in some edge cases).