tests: adding tests documenting behaviour from Issue #35
There seems to be some subtle bugs of sjasmplus involved, but as the
current documentation is under-specified in this regard, before fixing
anything in the code it would be better to design+document desired
behaviour precisely.

This test is just to archive the current state, including some comments
about it.