Fixing fix of Issue #33
This extra call to ReplaceDefine is redundant, ParseLine itself will
call the ReplaceDefine when the actual code is "running" (it is skipped
only when collecting DUP/REPT lines in non-running mode) ... hopefully
this time I understand it fully... probably not, but let's try it like
this for a while.

With the extra ReplaceDefine call, the block comments inside macro fail
(the "*/" gets parsed twice).