The manic miner src has a table generator with FOR/NEXT: can sjasmplus do it?
This was sort of challenge, if sjasmplus is capable to generate such
table in its current state (not having FOR/NEXT directives at all).

The first variant does use the DUP directive in manual way, recreating
the FOR/NEXT logic, but leading to quite complex and long code.

The second variant introduces `FOR` macro, which allow to define the
table with one-liner:

`FOR half, 0, $800, $800, <FOR y, 0, $E0, $20, <FOR x, 0, $700, $100, dw
ScreenBufferCWGI+x+y+half !> >`

Which seems to me reasonably close to original, although such `FOR`
macro imposes some finicky limits on the syntax and needs good
understanding of sjasmplus macro system to exploit its full potential,
but it's actually quite generic `FOR` for integer values.