test runner: new internal feature - custom *.cli files supported
this allows to run whole sub-script as custom test-runner for the
special tests which are impossible to cover by standard test runner.

This is internal undocumented feature, as the sub-script is being run as
part of main script, and it's not supposed to be used by ordinary tests,
only in special case when there is no way to write the test as standard
one, and review by Ped7g of such case is recommended.