Local labels refactored, fix #42
The new implementation is not using compiled-lines-counter any more, so
it is less vulnerable to amount of compiled lines changing between
different passes (different IF results, DUP counts, ALIGNs, ...)

Plus it should be a tiny bit faster... :D

While this will make issue #42 compile, it is sort of broken source in
principle, as the shown `IFNUSED` block will be always skipped in pass2
and pass3 (it may be assembled only in pass1, if ever). It's not clear,
what was author original intention, but it's highly likely there's
simpler way to write it in assembly, or it may be idea which requires
multi-pass assembler.