Fix of string literal parser to be more compatible with older sjasmplus
After the refactoring for v1.11.0 the expression ` 'a'|128 ` didn't
assemble any more, as the new logic was broken.

After this change expression like that will assemble as expected, but
`DC "a"-32` does treat the value as composed-expression, not string
literal, i.e. the result of such source will be value 65, not 65|128.

(this change in v1.11.0 behaviour did break the ZX Spectrum Next "NEX
loader" code in their firmware, while this fix does break only my own
test, which doesn't even make that much sense)