fix #45 - macro arguments are substituted more aggressively
There are some very good reasons why to not substitute term right after
ifdef/ifndef/define/defarray, but in case the term originates in macro
argument, this protection seems to be limiting the usage too much, so
this fix does lift that restriction (plus refactored part of the code
around).

This yields somewhat unlogical error messages if there is bug in the asm
source, redefining some define id twice, etc, so use with caution.