Fixed memory leak and line numbers in _lua_showerror()
The STRDUP'd LUA error message wasn't being freed in the end. It's unclear to me if this leak was the cause of fatal crashes on LUA errors, because for some reason it stopped crashing on my PC at this time (even older versions) and I can't reproduce it.

Also changed the way LUA line numbers are extracted. Previously it used a fixed offset +18 in an attempt to skip over the initial portion of error messages assuming that it is fixed. However, the initial portion of LUA error messages varies, e.g. path/filename for *.lua included files, or [string "script"] for inline LUA blocks within ASM. This caused the LUA errors to display wrongly at times, show incorrect line numbers and so on. Now it will try to look for a number enclosed within ":" instead, which seems to be the way line numbers are normally reported in LUA error messages.