env_var SJASMFLAGS renamed to SJASMPLUSOPTS
"SJASMFLAGS" seemed more like variable name from Makefile (which then
emits content of similar variables to the command line of the particular
tool, the tool itself does not use it), while this environment variable
is being detected and used by the assembler itself.

So rather renaming it to something tool-specific (similar to BASHOPTS,
SHELLOPTS, ...).