INCLUDE paths bugfix (no more "." added by default) (may break some projects!)
According to docs the INCLUDE should search relatively to the directory
of currently opened file (the one doing INCLUDE) (*and* on the include
paths provided on the command line or in environment variable).

If you have project like:
src/a.asm
src/b.asm

And you were assembling it with `sjasmplus src/a.asm` and `src/a.asm`
was doing `INCLUDE "src/b.asm"`, you need either fix the relative path
to `INCLUDE "b.asm"` or add `-I.` option: `sjasmplus -I. src/a.asm` to
give "src/b.asm" path chance to be found from that include path.