UNDEFINE: (!) removed undocumented ability to remove labels
Didn't seem to work properly, was undocumented and I was unable to make
sense of it.

As the DEFINEs have life-cycle per-pass, it makes sense to add/remove
them at particular line of source file.

But label's life-cycle is per-assembling (gets added in pass 1, and only
updated through pass 2 and 3), and removing it at certain line doesn't
make sense, as it will be removed for whole source in last pass, not
"after the UNDEFINE" line only.