CirrusCI: configuring windows builds (MinGW + VS 15 2017)
The tests are run only in the MinGW build, and their result is ignored
at this moment, as some tests fail due to slash vs backslash in filename
paths, and SHELLEXEC test suggests that the Win32 implementation is
maybe bugged.

The VS 15 build just exists and the resulting sjasmplus.exe is used only
to assemble examples, without further verification of results.