ctype.h: make sure the input `signed char` values will get converted to 0..255
The MS clib expects those `isdigit/.../is<char_trait>` functions to
receive `int` in range -1..255, i.e. valid char should be 0..255, but
rest of the code is often dereferencing (signed) char pointer, making
the VS build sometimes getting into trouble on UTF8 source files, etc.

So this is complete review of the source, casting the values to 0..255
in the same style everywhere, making the code base signed/unsigned char
agnostic.