tests: lighter output (lines+chars count) of main test-runner bash script
In an effort to see if the console I/O eats considerably time of windows
builds

But even in linux with 256+ tests it was becoming cumbersome to scroll
through.

Now any "diff" error is starting on previous console line which makes it
a bit more cryptic to decipher what's reported, but as long as all tests
does pass, the output is formatted well. (and error output is IMO still
reasonably formatted, just slightly tricky).