tests: added new test of defining/using structures across multiple files
Trying to hunt down the issue described by Peter G. on facebook, but so
far not successful, everything works as expected.

Waiting for further input from Peter to produce some example which fails
or explain the issue better. If I will be able to replicate it, I will
adjust the new test to show the ill behaviour, so the work on fix can
start.