revert INCLUDE paths bugfix. "." is added by default AGAIN. New "--inc" option
So the "fix" *did* break lot of projects, so the auto-addition of "." is
back.

But now the "--inc=path" option was extended, if used without the
"=path" part as simple "--inc", it will empty the path list completely.

If you really want to have in the list only directories explicitly
listed on command line, start (!) with --inc, like:
sjasmplus --inc --inc=path1 --inc=path2 file.asm