Fix ".end:" vs --dirbol, fix case-detection in cmphstr
"end:" and ".end:" should never work as END directive (they did when at
beginning of line and --dirbol option, the colon made them pass through
the test ... now any instruction string ending with colon is rejected as
instruction, as that case may happen only for labels).

"END" and ".END" should also not work as END directive with --dirbol
There was already test for these, but `cmphstr` function didn't compare
the upper-case keywords correctly when the keyword did start with dot
char (which is not the common usage scenario, but the ".end" test did
need it).

All added to tests + fixed.