User filenames in asm sources: convert backslash to slash on all platforms
This conversion is also visible in listing files/etc (!)

The regular slash should work also in Windows, but the backslash does
not work in Linux, so this should make windows-like sources easier to
assemble with linux sjasmplus binary.

(as the experience with windows devs shows, the filenames will have very
likely also wrong case, so the manual fixing of sources will be
inevitable. But if somebody has only wrong folder delimiter, this change
should cover it)

I'm a bit afraid this may actually break something... but I can't figure
out what it may be, so I'm enabling it now. Let's see what breaks...