error reporting refactored: LUA and MACRO errors show where they were defined
Reports the definition file+line if possible.

Still not perfect (would need to create unique "anonymous" names for
every inlined lua->endlua block, otherwise the error reporting has no
idea which "script" block was the origin one and just points to the
"endlua" directive causing emit of the error.

But it should not report completely non-existing lines any more, as it
was doing in previous version.