IF/IFN: forward reference is now only warning, and can be suppressed
I tried to write something smart to detect when IF/IFN block does emit
any byte, but the forward-reference flag is available only in pass1,
while the check of emit has to be done in every pass, so before
rewriting too much of sjasmplus, I gave up.

Instead the programmer can now explicitly suppress the message (demoted
from error to warning) with the end-of-line comment "ok".

If somebody will suppress it even if the code is incorrect, c'est la
vie.