tests refactoring + rebase + fixes of "addtrd" patch by Dart_Alver
Tests are fixed, using the new ".gz" bin-file compare feature.

Code is further fixed (needs fflush in early-exit scenarios on my linux
box)

Known bugs (to be fixed before merge):
- `&` "add" is not fortified against full catalog + non-existing file
name (will get out of `for` loop with bad fatPos)
- the error exits don't `fclose` and `fflush` where it would be
appropriate (when exiting because of value error, it surely should
cleanup, when exiting because fseek/fread/fwrite fails, then being
sloppy is probably ok, but still cleaner exit would be nice.
- rename "add" to "append"

Consider for refactoring: C++ streams instead (auto fclose/etc)? Read
TRD image into memory into structures, then rebuild it completely so it
will be also defragmented. I was avoiding this, but as the "&"
functionality already shuffles whole disc content, we are already there.
And the code is difficult to decipher (took me half a day to finally
understand what it does and how) and it works on 16B buffer. Not pretty.