LGTM: hide alert about unsafe function call
The call to `std::localtime` inside `main` should never become MT and
the content of shared buffer is immediately copied to own memory after
the call, so this is "false positive" in current code.

The proposed work-around is unfortunately "C2x", so disabling the alert
seems least harmful at this moment.