ELSEIF added, multi-ELSE warning removed, turned into error
The multi-ELSE block was funny curiosity before, but with introduction
of ELSEIF it does clash with the internal logic, so now always only
single conditional block can be active. The previous version was
alternating active/inactive on every ELSE, now after first active block
everything else (ELSE) is inactive, not alternating any more.

This is binary-change, but for very new feature which was not meant to
be actively used by users any way (and had warning).