parser: colon between end of EQU/DEFL expression and instruction is mandatory
The relaxed syntax without colon is ambiguous, line like:
```
label=$+1      and 7
```
was parsed as long expression, including the bitwise "and 7" into
calculation of symbol value (instead of parsing "and 7" as instruction).

With colon required it's now enforcing the correct style even on lines
where the relaxed syntax without colon did work before, and making the
sjasmplus syntax less ambiguous formally.