SLD: deprecating "F","D" types and adding new "L" type covering symbols
The "F","D" lines will be removed in future, the "L" should work as
complete replacement for them.

One benefit of "L" is adding clear boundaries to the symbol name, to
identify module/main/local parts of the symbol name, as it was difficult
to split the "F"/"D" data correctly.

Second benefit is adding "(usage) traits" info - how the symbol is
actually used and how is it defined. There's already some info provided,
but I hope to add more detailed traits info in the future versions of
sjasmplus.