Makefile: make all sjasm/*.(c|cpp) depend on all header files
to avoid [most of] problems when somebody is expecting the incremental
make builds to be correct when headers do change

(the Makefile of sjasmplus doesn't bother with correct c/cpp
dependencies based on header files, as usually GNU make is used for full
rebuilds)

This is still not fully correct/complete solution, but should cover most
of the practical cases.

Anyway, when in doubt (about binary), do `make clean` first.