fix #138: make the skipped conditional blocks more resilient to labels
The labels were causing the SkipFile() function to not find the
conditional block directive nested inside disabled block.

The fix is not perfect, works only when --dirbol is not used, and
doesn't work for temporary labels, but should be good enough for most
regular use-cases.

If you hit this issue with --dirbol or temporary label, just use two
lines please. The fully correct fix is not trivial, as the sjasmplus is
taking shortcuts to skip the disabled block in current architecture.