fix #164 invalid filenames in errors/warnings
On large projects with many includes the "file name archive" proved to
be wrongly implemented, invaliding c-string pointers which were supposed
to be "permanent", causing error reporting to print bogus filename.

I was unable to reproduce the issue in individual test file, but the
behavior happened in https://github.com/TinfoilAsteroid/EliteNext
project and I was able to reproduce it with it.