internals: refactoring source/macro/error line numbering implementation
making the system more powerful keeping track of full include depth and
enable unlimited amount of sub-emits on top of that.

This is in preparation for fix of lua error reporting in complex cases
and also to enable more verbose error output including "included from"
in future.

This is the first part of refactoring, keeping the visible output
changes to minimum:
- DUP emits into SLD only source position, not duplicating definition
position along (removed unplanned redundancy? almost fix)
- macro in macro reports more correctly where error happened and where
it was emitted (fix)
- macro error inside another macro reports "emitted" only one level up,
not all way into source (change)
- while infinite loop points to the WHILE source line (fix)
- relocation start error tuned to be "here"+"emitted" double message
(change)