cli: include options now accept paths in next argument, enabling ~ substitution
old syntax "-I~/path/" -> this is ignored by the shell, so "~" is not
substituted with $HOME, because it's one string

new syntax: "-I ~/path/" -> shell will treat ~/path as individual string
and substitute the ~ (if it does that usually for you)

Also the sjasmplus will now check all include paths before starting
assembling and report if some can't be opened.

And it reports if some starts with literal "~" (fopen will not
substitute that, so if shell didn't before starting sjasmplus, report it
as error).

OLD SYNTAX STILL WORKS (and it's not going to be removed any time soon,
if ever), but docs and --help will suggest the new syntax only.