c++17: replacing custom code with std::filesystem[::path] where feasible
Adding GetInputFileName using std::filesystem::path and making it work
for main OpenFile and IncludeFile.

TODO:
- convert other input file situations like INCBIN, etc..
- get rid of obsolete functions later

Committing this as intermittent step just in case I will be unable
to develop this further soon or if I need rollback to working version.
(and to verify all CI platforms to pass tests)